+++ tests/resources/testdata/TestPortalCustomization/CRUD/2/main-step-2.tf
@@ -5,12 +5,4 @@ resource "konnect_portal" "my_portal_for_customization" {
 
 resource "konnect_portal_customization" "my_portal_customization" {
   portal_id = konnect_portal.my_portal_for_customization.id
-  robots    = "my_robots"
-  theme = {
-    colors = {
-      primary = "#000000"
-    }
-    mode = "system"
-    name = "my_name"
-  }
 }